+++ README.md
@@ -1,0 +1,3 @@
+# Battleship
+
+A simple python implementation of the famous battleship game
